requirements.txt
# LangChain Core Dependencies
langchain>=0.1.0
langchain-community>=0.0.20
langchain-groq>=0.1.0
langchain-classic>=1.0.0
langchain-text-splitters>=0.1.0
langchain-huggingface>=0.0.1

# Vector Store
faiss-cpu>=1.7.4

# Embeddings
sentence-transformers>=2.2.0
torch>=2.0.0

# Environment Variables
python-dotenv>=1.0.0

# Additional dependencies that may be required
numpy>=1.24.0
pydantic>=2.0.0

# Web Framework
streamlit>=1.28.0
